refactor(app): extract the result-grid interaction wiring
The left pane's client-side grid handlers — global filter, per-column
filter row, sort-by-header, column drag-reorder, and the Columns
visibility popup — move to `wire::grid`. They all re-derive the shown
grid from the cached base view and never touch the driver, so they group
cleanly.

Per-pane handles are taken off `AppState::panes` inside the module
instead of being aliased into locals up in `main`.
