+++ app/src/wire/mod.rs
@@ -5,5 +5,6 @@
 //! and calls them in order; the handler bodies themselves are unchanged.
 
 pub(crate) mod conn_form;
+pub(crate) mod grid;
 pub(crate) mod settings;
 pub(crate) mod update;
